+++ Editor/AssemblyInfo.cs
@@ -1,8 +1,8 @@
 ﻿using System.Reflection;
 using System.Runtime.CompilerServices;
 
-[assembly: AssemblyVersion("2.2.4.0")]
-[assembly: AssemblyFileVersion("2.2.4.0")]
-[assembly: AssemblyInformationalVersion("2.2.4")]
+[assembly: AssemblyVersion("2.3.1.0")]
+[assembly: AssemblyFileVersion("2.3.1.0")]
+[assembly: AssemblyInformationalVersion("2.3.1")]
 
 [assembly: InternalsVisibleTo("YarnSpinner.Editor.Tests")]