+++ Runtime/Views/OptionView.cs
@@ -50,6 +50,13 @@ namespace Yarn.Unity
 
         public void InvokeOptionSelected()
         {
+            // turns out that Selectable subclasses aren't intrinsically interactive/non-interactive
+            // based on their canvasgroup, you still need to check at the moment of interaction
+            if (!IsInteractable())
+            {
+                return;
+            }
+            
             // We only want to invoke this once, because it's an error to
             // submit an option when the Dialogue Runner isn't expecting it. To
             // prevent this, we'll only invoke this if the flag hasn't been cleared already.