+++ Runtime/YarnProject.cs
@@ -224,6 +224,17 @@ namespace Yarn.Unity
                 return cachedProgram;
             }
         }
+        private void Awake() 
+        {
+            // We have to invalidate the cache on Awake.
+            // Note that this cannot be done through the importer 
+            // (e.g., with a setter method that sets compiledYarnProgram and invalidates cachedProgram)
+            // because the YarnProject the importer accesses is NOT the same object as the 
+            // one currently loaded in the editor. (You can tell by comparing their HashCodes)
+            // If there are other sources that can change the value of compiledYarnProgram aside from
+            // the importer in runtime, maybe we can add such a method, but until then, this is sufficient.
+            cachedProgram = null;
+        }
     }
 
     public enum LocalizationType {